+++ packages/core/src/ceramic.ts
@@ -223,6 +223,7 @@ export class Ceramic implements CeramicApi {
       this._streamHandlers
     )
     const pinStore = modules.pinStoreFactory.createPinStore()
+    const localIndex = new LocalIndexApi(modules.indexing, this.repository, this._logger)
     this.repository.setDeps({
       dispatcher: this.dispatcher,
       pinStore: pinStore,
@@ -230,8 +231,10 @@ export class Ceramic implements CeramicApi {
       handlers: this._streamHandlers,
       anchorService: modules.anchorService,
       conflictResolution: conflictResolution,
+      indexing: localIndex,
     })
-    this._index = new LocalIndexApi(modules.indexing, this.repository, this._logger)
+    // TODO (NET-1687): remove indexing part of refactor to push indexing into state-manager.ts
+    this._index = localIndex
   }
 
   get index(): IndexApi {
@@ -625,38 +628,9 @@ export class Ceramic implements CeramicApi {
       this.repository.updates$
     )
 
-    await this.indexStreamIfNeeded(stream)
+    await this.repository.indexStreamIfNeeded(state$)
 
     return stream
-  }
-
-  /**
-   * Helper function to add stream to db index if it has a 'model' in its metadata.
-   * @param stream
-   * @private
-   */
-  private async indexStreamIfNeeded(stream: Stream): Promise<void> {
-    if (!stream.metadata.model) {
-      return
-    }
-
-    const asDate = (unixTimestamp: number | null | undefined) => {
-      return unixTimestamp ? new Date(unixTimestamp * 1000) : null
-    }
-
-    // TODO(NET-1614) Test that the timestamps are correctly passed to the Index API.
-    const lastAnchor = asDate(stream.state.anchorProof?.blockTimestamp)
-    const firstAnchor = asDate(
-      stream.state.log.find((log) => log.type == CommitType.ANCHOR)?.timestamp
-    )
-    const STREAM_CONTENT = {
-      model: stream.metadata.model,
-      streamID: stream.id,
-      controller: stream.metadata.controller,
-      lastAnchor: lastAnchor,
-      firstAnchor: firstAnchor,
-    }
-    await this._index.indexStream(STREAM_CONTENT)
   }
 
   /**
@@ -695,7 +669,7 @@ export class Ceramic implements CeramicApi {
       this.repository.updates$
     )
 
-    await this.indexStreamIfNeeded(stream)
+    await this.repository.indexStreamIfNeeded(state$)
 
     return stream
   }