+++ packages/3id-did-resolver/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.5-rc.1](/compare/@ceramicnetwork/3id-did-resolver@2.1.5-rc.0...@ceramicnetwork/3id-did-resolver@2.1.5-rc.1) (2022-08-11)
+
+**Note:** Version bump only for package @ceramicnetwork/3id-did-resolver
+
+
+
+
+
+## [2.1.5-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/3id-did-resolver@2.1.4...@ceramicnetwork/3id-did-resolver@2.1.5-rc.0) (2022-08-10)
+
+**Note:** Version bump only for package @ceramicnetwork/3id-did-resolver
+
+
+
+
+
 ## [2.1.4](/compare/@ceramicnetwork/3id-did-resolver@2.1.4-rc.2...@ceramicnetwork/3id-did-resolver@2.1.4) (2022-08-08)
 
 **Note:** Version bump only for package @ceramicnetwork/3id-did-resolver