+++ packages/ipfs-daemon/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2-rc.1](/compare/@ceramicnetwork/ipfs-daemon@2.1.2-rc.0...@ceramicnetwork/ipfs-daemon@2.1.2-rc.1) (2022-08-11)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-daemon
+
+
+
+
+
+## [2.1.2-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-daemon@2.1.1...@ceramicnetwork/ipfs-daemon@2.1.2-rc.0) (2022-08-10)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-daemon
+
+
+
+
+
 ## [2.1.1](/compare/@ceramicnetwork/ipfs-daemon@2.1.1-rc.2...@ceramicnetwork/ipfs-daemon@2.1.1) (2022-08-08)
 
 **Note:** Version bump only for package @ceramicnetwork/ipfs-daemon